functions/main.py
"""
Firebase Functions (Python) - Alpaca Account Management
These functions handle Alpaca API integration which requires Python SDK
"""

import json
import os
from datetime import datetime
from firebase_functions import https_fn, options
from firebase_admin import initialize_app, auth, db
import alpaca_trade_api as tradeapi
from alpaca.broker.client import BrokerClient
from alpaca.broker.models import Account
from alpaca.broker.requests import CreateAccountRequest
from alpaca.broker.enums import AccountType

# Initialize Firebase Admin
app = initialize_app()

# Alpaca Configuration
ALPACA_API_KEY = os.environ.get('ALPACA_API_KEY')
ALPACA_SECRET_KEY = os.environ.get('ALPACA_SECRET_KEY')
ALPACA_BROKER_API_KEY = os.environ.get('ALPACA_BROKER_API_KEY')
ALPACA_BROKER_SECRET = os.environ.get('ALPACA_BROKER_SECRET')

# Initialize Alpaca Broker Client for paper trading accounts
if ALPACA_BROKER_API_KEY and ALPACA_BROKER_SECRET:
    broker_client = BrokerClient(
        api_key=ALPACA_BROKER_API_KEY,
        secret_key=ALPACA_BROKER_SECRET,
        sandbox=True  # Use sandbox for paper trading
    )
else:
    broker_client = None
    print("Warning: Alpaca Broker API credentials not configured")


@https_fn.on_request(
    cors=options.CorsOptions(
        cors_origins="*",
        cors_methods=["POST", "OPTIONS"],
    )
)
def createAccount(req: https_fn.Request) -> https_fn.Response:
    """
    Create Alpaca paper trading account
    Matches original Cloud Function endpoint
    """

    # Handle CORS preflight
    if req.method == 'OPTIONS':
        return https_fn.Response('', status=204)

    # Verify Firebase Auth
    auth_header = req.headers.get('Authorization', '')
    if not auth_header.startswith('Bearer '):
        return https_fn.Response(
            json.dumps({'error': 'Unauthorized'}),
            status=401,
            headers={'Content-Type': 'application/json'}
        )

    try:
        # Verify the Firebase ID token
        id_token = auth_header.split('Bearer ')[1]
        decoded_token = auth.verify_id_token(id_token)
        user_id = decoded_token['uid']

        # Parse request body
        request_json = req.get_json(silent=True)
        if not request_json or 'agentId' not in request_json:
            return https_fn.Response(
                json.dumps({'error': 'Agent ID is required'}),
                status=400,
                headers={'Content-Type': 'application/json'}
            )

        agent_id = request_json['agentId']

        # Check if account already exists in Firebase
        account_ref = db.reference(f'users/{user_id}/accounts/{agent_id}')
        existing_account = account_ref.get()

        if existing_account and existing_account.get('accountId'):
            return https_fn.Response(
                json.dumps({
                    'success': True,
                    'accountId': existing_account['accountId'],
                    'message': 'Account already exists'
                }),
                headers={'Content-Type': 'application/json'}
            )

        # Create Alpaca paper trading account
        if broker_client:
            try:
                # Create a sub-account for paper trading
                account_request = CreateAccountRequest(
                    contact={
                        "email_address": f"{user_id}-{agent_id}@spookylabs.com",
                        "phone_number": "555-666-7788",
                        "street_address": ["20 N San Mateo Dr"],
                        "city": "San Mateo",
                        "state": "CA",
                        "postal_code": "94401"
                    },
                    identity={
                        "given_name": "Agent",
                        "family_name": agent_id[:8],
                        "date_of_birth": "1990-01-01",
                        "tax_id_type": "SSN",
                        "tax_id": "666-55-4321",
                        "country_of_citizenship": "USA",
                        "country_of_birth": "USA",
                        "country_of_tax_residence": "USA",
                        "funding_source": ["employment_income"]
                    },
                    disclosures={
                        "is_control_person": False,
                        "is_affiliated_exchange_or_finra": False,
                        "is_politically_exposed": False,
                        "immediate_family_exposed": False
                    },
                    agreements=[
                        {
                            "agreement": "customer_agreement",
                            "signed_at": datetime.utcnow().isoformat() + "Z",
                            "ip_address": "127.0.0.1"
                        }
                    ],
                    enabled_assets=["us_equity"],
                    account_type=AccountType.PAPER
                )

                # Create the account
                account = broker_client.create_account(account_request)
                account_id = account.id

            except Exception as alpaca_error:
                print(f"Alpaca API error: {str(alpaca_error)}")
                # Fall back to mock account for development
                account_id = f"mock-{agent_id[:8]}"
        else:
            # Use mock account if Alpaca not configured
            account_id = f"mock-{agent_id[:8]}"

        # Save account info to Firebase
        account_ref.set({
            'accountId': account_id,
            'agentId': agent_id,
            'userId': user_id,
            'createdAt': datetime.utcnow().isoformat(),
            'status': 'created',
            'funded': False
        })

        return https_fn.Response(
            json.dumps({
                'success': True,
                'accountId': account_id,
                'message': 'Account created successfully'
            }),
            headers={'Content-Type': 'application/json'}
        )

    except auth.InvalidIdTokenError:
        return https_fn.Response(
            json.dumps({'error': 'Invalid authentication token'}),
            status=401,
            headers={'Content-Type': 'application/json'}
        )
    except Exception as error:
        print(f"Error creating account: {str(error)}")
        return https_fn.Response(
            json.dumps({'error': 'Internal server error'}),
            status=500,
            headers={'Content-Type': 'application/json'}
        )


@https_fn.on_request(
    cors=options.CorsOptions(
        cors_origins="*",
        cors_methods=["POST", "OPTIONS"],
    )
)
def fundAccount(req: https_fn.Request) -> https_fn.Response:
    """
    Fund Alpaca paper trading account
    Matches original Cloud Function endpoint
    """

    # Handle CORS preflight
    if req.method == 'OPTIONS':
        return https_fn.Response('', status=204)

    # Verify Firebase Auth
    auth_header = req.headers.get('Authorization', '')
    if not auth_header.startswith('Bearer '):
        return https_fn.Response(
            json.dumps({'error': 'Unauthorized'}),
            status=401,
            headers={'Content-Type': 'application/json'}
        )

    try:
        # Verify the Firebase ID token
        id_token = auth_header.split('Bearer ')[1]
        decoded_token = auth.verify_id_token(id_token)
        user_id = decoded_token['uid']

        # Parse request body
        request_json = req.get_json(silent=True)
        if not request_json:
            return https_fn.Response(
                json.dumps({'error': 'Invalid request body'}),
                status=400,
                headers={'Content-Type': 'application/json'}
            )

        agent_id = request_json.get('agentId')
        account_id = request_json.get('accountId')
        amount = request_json.get('amount', 100000)  # Default $100k

        if not agent_id or not account_id:
            return https_fn.Response(
                json.dumps({'error': 'Agent ID and Account ID are required'}),
                status=400,
                headers={'Content-Type': 'application/json'}
            )

        # Verify account ownership
        account_ref = db.reference(f'users/{user_id}/accounts/{agent_id}')
        account_data = account_ref.get()

        if not account_data:
            return https_fn.Response(
                json.dumps({'error': 'Account not found'}),
                status=404,
                headers={'Content-Type': 'application/json'}
            )

        if account_data.get('accountId') != account_id:
            return https_fn.Response(
                json.dumps({'error': 'Account mismatch'}),
                status=403,
                headers={'Content-Type': 'application/json'}
            )

        # Check if already funded
        if account_data.get('funded'):
            return https_fn.Response(
                json.dumps({
                    'success': True,
                    'message': 'Account already funded',
                    'balance': amount
                }),
                headers={'Content-Type': 'application/json'}
            )

        # For paper trading accounts, funding is automatic
        # Real Alpaca paper accounts start with $100k by default

        if broker_client and not account_id.startswith('mock-'):
            try:
                # Verify account exists and is active
                account = broker_client.get_account_by_id(account_id)

                # Paper accounts are automatically funded
                # Just verify the account is in good standing
                if account.status != 'ACTIVE':
                    # Wait for account approval (paper accounts are usually instant)
                    pass

            except Exception as alpaca_error:
                print(f"Alpaca verification error: {str(alpaca_error)}")
                # Continue anyway for paper trading

        # Update account in Firebase
        account_ref.update({
            'funded': True,
            'fundedAt': datetime.utcnow().isoformat(),
            'initialBalance': amount,
            'currentBalance': amount
        })

        return https_fn.Response(
            json.dumps({
                'success': True,
                'message': 'Account funded successfully',
                'balance': amount
            }),
            headers={'Content-Type': 'application/json'}
        )

    except auth.InvalidIdTokenError:
        return https_fn.Response(
            json.dumps({'error': 'Invalid authentication token'}),
            status=401,
            headers={'Content-Type': 'application/json'}
        )
    except Exception as error:
        print(f"Error funding account: {str(error)}")
        return https_fn.Response(
            json.dumps({'error': 'Internal server error'}),
            status=500,
            headers={'Content-Type': 'application/json'}
        )